requirements.txt
yfinance
pandas
pandas-datareader
confluent-kafka
python-dateutil
numpy
pymysql
elasticsearch[async]==8.5
pyspark